code: edited TableManagerTemplate to be flexible for multiple entities, added TableManagerTemplate subclasses for City and Stopover, Edited LoginPage to redirect to MainMenu, MainMenu overhauled
In TableManagerTemplate, getAllRecords, Add, Delete, and Update were modified to allow flexibility for multiple entities. Some helper functions were added to provide this flexibility.

CityManagerTable extending TableManagerTemplate was added. Functions that override some functions in the TableManagerTemplate superclass are present in order to fit the entity's requirements.

StopoverManagerTable extending TableManagerTemplate was added. Functions that override some functions in the TableManagerTemplate superclass are present in order to fit the entity's requirements.

MainMenu was edited so that buttons direct to each entity's specific TableManager. LoginPage now directs to MainMenu.

See Changelog on GoogleDocs for full detailed list of changes