clean: added documentation to TableManagerTemplate and removed deprecated packages fix: corrected error in TableManagerTemplate's getRecord() code: added helper function to getRecord()
Documentation to TableManagerTemplate added to further clarify the purpose
of each of each function and reduce confusion for later use. Documentation
for this class is now complete.

Packages that were no longer being used were delete to reduce clutter in
the workspace. Consider refactoring the remaining classes into separate packages
for cleanliness.

getRecord() had an error which resulted in the function using a primary key
of a table to retrieve a record in a different table. This has been correceted.

Helper function displayRecord() added to help getRecord(). Function works correctly
and now must be overwritten by subclasses to be tailored for its use.